Unify pipette and clipboard
